requirements.txt
asgiref==3.5.2
autopep8==1.7.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-authemail==2.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.14
whitenoise==6.2.0
